Make a symbol subclass of string for WORD! spellings
The term "REBSYM" had been used for the symbol ID enumeration, but that
name was desired for a subclass of REBSTR that has been checked as
valid for a wordspelling.

This changes the former REBSYM to "SYMID", and makes the subclass.

The change identified a problem in that the filename mechanics (for
storing a filename and line number in blocks to say where they came
from) was previously lying and saying that it was interned as a valid
word, when filenames contain invalid word characters (like / and .)

So this moves that to using plain strings, and just freezes them to
ensure the user does not change them after passing the filename in.
(Copying would be slow and wasteful; the best idea would be for users
to have the responsibility of making their own copies.)