Make contexts/actions/maps nodes, fix more strict alias
Turning up optimizations in the compiler gave warnings regarding the
type checking technique used for `mutable_LINK` and `mutable_MISC`.

This generalizes the approach to a method that does not violate
strict aliasing, but only has teeth in C++.  It establishes some
"ensure" macros that invoke templated classes to check the type.

In addition, this uses multiple inheritance to solve the problem of
making REBCTX*, REBMAP*, and REBACT* be REBNOD* without a cast...
while still not letting them act like their base class REBSER*
or REBARR*.  The result is less need for NOD() casts.