Reduce duplicate code paths in DEBUG_UTF8_EVERYWHERE
The initial formulation of the REBCHR(*) type predated the const_if_c
technique, as well as the ability to m_cast() a REBCHR(const*) to
a REBCHR(*).  With those two methods available, we can force the
UTF-8 code through a common code path, which gets rid of redundancy
and makes it easier to work on optimizations.