Some work switching from NOT_END() to tail checks
It is now planned to move away from depending on array terminators,
and instead comparing against the tail of arrays during enumeration.

This makes a start on the idea, by adding parameters to the bind
routines, and working through some other files as well.  It's a long
and boring task...so breaking it up across several commits interspersed
with other work is probably the best way to go about it.