Restyle "enclose specialize" as "specialize enclose"
There was an idiom of setting a parameter to a dummy value via
specialization to remove it from a function's interface, and then
overwriting it in an enclosure.  However, with more strict enforcement
of the idea that specialized out parameters "don't exist" in the
interface, there is no reason the enclosure should be able to see
them any more than a local.

An easy remedy for this is to do the enclosure first (while the
parameter is still present) and then specialize the result of that.
This makes POINTFREE and the inner draft implementation of multiple
returns use the updated method.