More efficient DO of FRAME!
The DO code wanted to avoid multiple DO of the same frame, since the
variables in that frame are corrupted by function execution (a function
can freely modify its inputs).

Previously, to achieve this it would "steal" the frame data out from
under the varlist node...leaving the former node as an inaccessible
stub (so outstanding references would not crash).  Then it would
allocate a new frame identity to reuse the data for the actual
execution.

This commit tries a more efficient approach, to consider a DO-able
frame to be one that is "phaseless"...and have a single bit on a frame
say whether it has been executed before or not.  If a phaseless frame
tries to DO a varlist that has already been invoked, that can be an
error even while keeping the frame itself accessable when viewed
through the lens of the executing phases.