Fix MAXIMUM, MINIMUM
The Compare() functions in R3-Alpha used a more confusing scheme for
return codes...which were simplified to a tristate of `-1, 0, or 1` for
the base case.

Apparently the MINIMUM and MAXIMUM functions were wired backwards, and
there was no test to catch it.  This fixes it and adds a basic test
for integers 0 and 1.