Move the two GC-mark bits into the node byte
Both REBVAL and REBSER have in common that they use two bits in the
header to tell the GC if that specific instance of the value has
nodes that need marking in them.

R3-Alpha lacked these bits--which forced the GC to run a switch()
statement on the type of values.  And series marking was not as
generalized.

Since these two bits exist for both categories of values, it is
not particularly forced to give them aliases in the node byte...thus
freeing up two CELL_FLAGs and two SERIES_FLAGs for other usages.