Remove variadic behavior from DOES
An experimental change to DOES made it quote non-block WORD!/PATH!
arguments and implement a "does by example" shorthand:

    >> block: []

    >> ap: does append block <demo>

    >> ap
    == [<demo>]

    >> ap
    == [<demo> <demo>]

This behavior caused the implementation of DOES to be more complex, and
also gave it different properties.  For instance, in the example above
a change to block will not change what the AP appends to, while it
would change if you'd said `does [append block <demo>]` because it
would run the block each time (vs. a specialization).

Today, there is a more general tool for accomplishing such goals...
the REFRAMER.  It can express the above as:

    does+: reframer func [f [frame!]] [
        does [do copy f]
    ]

Also the "weird" nature of this impromptu specialization is a better
fit for generators like POINTFREE (<-).