Remove variadic/quoting interface from MATCH
When it was noticed that MATCH could take filtering functions, the
question of how to use a filter function that took more than one
parameter came up.  At the time, PARSE returned a LOGIC!, and so
the idea of being able to have MATCH expressed with the example call
and "steal" its first parameter came up:

   >> match parse "aaa" [some "a"]
   == "aaa"

   >> match parse "aaa" [some "b"]
   ; null

The mechanics behind this trick evolved over time into the REFRAMER
native, which allows functions with this "odd" behavior to be made.
For instance, this case could be written as:

    match+: reframer func [f [frame!] <local> p] [
        p: f/(first parameters of action of f)
        if do f [p]
    ]

So removing the convoluted code from the MATCH native makes sense, as
the functionality is a bit dicey (forces quoting parameter convention
for example).  It helps put focus on optimizing and enhancing the
more generalized REFRAMER native.

In addition, the most compelling example of this feature was with
PARSE to get the initial input.  That's been changed to the default
behavior, so it's less critical to have an optimized way to do it
with MATCH-the-native.