Allow DO of QUOTED!
Rather than try and do bizarre ways of optimizing specializations of
the IDENTITY function with an ACTION!, it seems easier to make
QUOTED! able to fit anywhere you could use an action...simply to
manifest itself as being a very efficient zero-arity action that
returns the value it contains.

Making the foundational DO operation take QUOTED! is a step in this
direction.