Make REFRAMER pass QUOTED! if not an action
This consolidates the code shared between MAKE FRAME! of VARARGS! and
REFRAMER a bit more.  It also makes it possible to get a QUOTED! as
the argument to the reframer, and gives a FRAME! for IDENTITY if you
make frame for a varargs that isn't a function.

(These features aren't really ready for prime time, and the focus is
on paring down the total amount of code, so this is untested for now.)