Eliminate SYM__ANONYMOUS__
The general rule has become that `nullptr` is used for slots where
no label is required.  The --anonymous-- label is a holdover from
where te currency was cells (not series) and hence a fully different
datatype would need to be used (e.g. NONE!/BLANK!) which was more
of an issue than when pointers to REBSYM/REBSTR are used.

This switches to using nullptr (#define'd as ANONYMOUS)