Stale FRAME!s raise errors
With the memory-and-time saving choice to not "steal" context variables
when DO FRAME! is executed, there still needs to be a way to notice
that a frame has been handed to execution...and should not be re-done
again, nor should the frame values be assumed to be caches of the
inputs.

This adds an error to catch that situation, which is based on a bit
that the evaluator puts on when actions are begun.