Update `x_cast` concept
Because they aren't inlined out of debug builds, DEBUG_CHECK_CASTS
makes the `cast()` operator slow.  With vetted operators that know
more or less what they are doing, this is wasted CPU time.

But C-style casts as parentheses are still hard to see.  So x_cast()
was introduced to not be a checked cast form, while still having a
more visible syntax.

It was thought that this cast form could cast away const, but it turns
out that C++ optimized builds were complaining about this when non
optimized builds were not.  This means some forms that had been
changed to use just x_cast instead of a cast to a const form and a
mutable form wouldn't work.  So those are fixed here.