Fix STATS/PROFILE, get rid of CLEAR()/CLEARS()
The statistics object wasn't being zeroed out in startup.  This does
that, as well as gets rid of the CLEAR() and CLEARS() macros that are
really just obfuscations for memset() of 0, forcing C programmers
to look up what it does (when they're perfectly capable of reading
the other calls to memmove() and memcmp()...)

Also converts the stats to use the API fully for the object creation
(it had been half-converted at some point), and adds OBJECT! to the
return type spec.