Control debug stats collection with separate switch
This makes it possible to ask for statistics about series creation
and memory use to be collected in a non-"Debug" build by setting the
separate flag DEBUG_COLLECT_STATS.

It also adds that flag to the emscripten build, to do a comparison and
see what the memory burden is vs. the desktop build.