Get rid of MAKE ACTION! on a BLOCK!
R3-Alpha had a concept that the fundamental operation for making
actions went through MAKE.  The way to get two parameters to MAKE was
to pass a spec block and a body block via a 2-element outer block.

This somewhat convoluted idea makes little sense, compared to calling
the entry point FUNC and having it be an arity-2 function.

Ren-C's strategy is to build more fundamental operations for action
construction via the FRAME! interface creation, and then being able
to ADAPT blocks into frames...transforming the results back and forth
between FRAME! and ACTION!.  This has shown enough promise to be
certain that the MAKE ACTION! synonym for FUNC is ultimately a
confusing and inefficient concept...that makes the spec dialect seem
more baked in than it should be.