Eliminate "pseudotype" for faster QUOTED! check
The scarcity of bits on parameters had forced an undesirable use of
`REB_P_XXX` "pseudotypes" > REB_MAX.  This convoluted the test for
if things were quoted, and was confusing in general.

With keylists broken out into separate series, the "PARAM" was moved
toward greater simplicity and could be a typeset with extra bits
where the symbol used to be.  That killed off pseudotypes.

This reaps the benefit of that by making the check for quoted simpler.