Avoid undefined behavior in MATCH
The experimental MATCH dialect needs to fetch values, as well as
potentially peform evaluations while holding those values stable
across the evaluations.

In order to avoid the cost of a temporary GC guarded cell, the code
tried to reuse the known-to-be-GC-guarded output cell to hold the
fetched test until the result was needed.  However, expressions that
initialized the output while using the input could run into trouble:

    Get_If_Word_Or_Path_Throws(out, ...);
    test = out;
    ...
    Init_Logic(out, Some_Operation(test, ...));

The problem comes in when these boil down to macros, and there is no
guaranteed sequence for the compiler to evaluate the expression.

This goes ahead and re-styles the function to guard a temporary cell
on each recursion, with all exit paths dropping the guard.