Fixes for mutable_BINDING() ub sanitizer issues
Before a trick was figured out for REBCTX* and REBACT* to derive
from REBNOD in the C++ build, the assignment of them to REBNOD*
variables (or REBSER/REBARR) required extraction of their main
node, e.g. CTX_VARLIST() or ACT_DETAILS().  This was in the moment
seen as more "proper" than using the NOD() cast operation.

However, CTX_VARLIST() and ACT_DETAILS() would dereference null
pointers if passed...while the cast operators would not.  So that
was bad in the case of bindings which could be null.

But good news is that with being subclasses of NOD(), no cast is
necessary anymore, and they can be assigned directly...allowing
nulls as well.