Fix bug with DEBUG_TRACK_CELLS and optimization
Debug builds would attempt to format cells with information regarding
the file, line, and tick where they were initialized.  Capturing such
information has to be done with macros...so each init routine was
done as a macro that would apply tracking to the output cell before
calling the "real" initialization:

    #define Init_Logic(out,flag) \
        Init_Logic_Core(TRACK_CELL_IF_DEBUG(out), (flag))

This strategy has the advantage that in debug builds (which do not
do inlining) it would have no cost if the tracking feature wasn't used.
But importantly, it's lightweight in terms of not needing to tunnel
the file/line information through to the core function...which would
multiply the #ifdefs in an unpleasant way.

BUT...this had a problem, because C (and C++) do not guarantee the
order in which parameters to a function are evaluated.  So if you have
a callsite that says something like:

    Init_Logic(D_OUT, VAL_LOGIC(D_OUT));

That turns into:

    Init_Logic(TRACK_CELL_IF_DEBUG(D_OUT), VAL_LOGIC(D_OUT));

So the tracking information could be added before the value to
initialize was extracted.  This means that if that information overlaps
the payload (e.g. where the logic TRUE/FALSE bit lives), the VAL_LOGIC
call could read corrupted data.

That means only one of the two modes that debug tracking of cells were
using can work: the one that extended the cell size.  However, that
was not the default...the default would just use whatever cell space
that wasn't overwritten by cell payloads.

This commit reduces the debug track mode to only use the extended form,
so that this pattern can be used in the Init_XXX() macros without
concern about the reordering.  In practice, the "cheap" form of debug
tracking that didn't extend cells was barely useful anyway...and was
probably not worth the time it took up since most cells overwrote the
data anyway.

So it's both simplifying and plugs a correctness problem.