Forbid CELL_KIND() on REB_QUOTED
The concept behind REBCEL(const*) is that it is supposed to be the
type of an "unescaped" cell.  So it's either a view of a cell with 3
levels of quoting or less under the perspective that it is not quoted,
or the payload of the REB_QUOTED which has no levels of quoting.

But making `const RELVAL` not coerce to REBCEL implicitly would be
inconvenient.  Routines that take REBCEL are designed to assume
the type of the cell is already known (e.g. VAL_WORD_SPELLING() takes
a REBCEL and asserts it is a word, while IS_WORD() won't take REBCEL)

This tweaks a couple of routines that weren't in line with the rules,
and makes IS_RELATIVE() operate only on RELVAL()--as a REBCEL() is
not supposed to have a binding.