Get rid of `uint_fast32_t` header field
There was a half-baked idea to get the header size laid out with
`uintptr_t`, but actually use `uint_fast32_t`.  However, the actual
byte-order-sensitive macros were all set up to put the flags in the
positions that would work with `uintptr_t`.  So if `uint_fast_32_t`
was a different size, the positions would be wrong.

Fixing that would be complex.  But furthermore, even if it could be
made to work it would then rule out use of the extra space on a 64-bit
platform for optimization purposes...because of strict aliasing.

This gets rid of that complexity, and does a bit of other light reorg
and commenting.

Also notices that REBFLGS was uint_fast32_t and holding byte-order
sensitive flags, so changes that to uintptr_t as well...with a note
about not using it with these flags if it's expected to be a different
datatype.