Find Patch Variant List from Context
Although bits are in short supply inside a Context VARLIST's REBSER
node, this makes the MISC() field do double duty as either the
meta or the most recently-created patch variant for that context.

If it's a patch variant, then the meta is moved to take up the MISC()
slot of the patch.  This has two suboptimal consequences:

* The most recently-created patch variant of a context will not be
  GC'd, because it has to be marked GC-live in order to get it to
  mark its meta.

* All patches give up their MISC() field for other uses, just so that
  one of them can hold the meta for the context.

While that's a bit annoying, over the long run it's likely more
optimal than trying to create a global hash table that maps from
contexts to patch lists.  The strategy of doing an update after every
creation so it only keeps the most recently created patch variant live
makes it likely faster to get a useful result and also makes it more
likely to be able to GC down to just one.

The code is reorganized to have just one function for creating patches
instead of the two that existed before, and the name makes it clear
that it will "Make_Or_Reuse_Patch()".