Primitive PARSE/INSIDE Virtual Binding Mechanism
This allows you to specify a single object that will be used to
virtually bind the BLOCK! rules and GROUP! code for a PARSE.

It is an experiment to see how well such a mechanism works.  Better
would be if the binding could be applied selectively as part of
a rule during a parse, or to be able to apply to just BLOCK!s or
just GROUP!s.