Fix stale group evaluation bit
There was a bug where groups that ended with a stale evaluation were
not clearing the stale bit.

Using the version of evaluation that leaves the stale bit on appears
to have been done on purpose at some point.  But leaving it that way
when the evaluation finishes leads to bugs, e.g. `any [(10 elide "hi")]`
would evaluate to NULL.

This leaves the stale operation but comments that it's not clear why
it's there, and clears the stale bit.