Move Dynamic/Singular Status To `leader` flags
This makes some moves designed to ultimately be able to claim some a
pointer-sized slot out of the `info` portion of a series node.  It
starts in that direction by not using the status of the LEN and WIDE
bytes in info for arrays.

To determine if a singular-allocated array contains 0 or 1 item, it
is tested for END marker status.

It effectively frees the LEN byte for both singular and non-singular
arrays... and the LEN byte in dynamic non-arrays  (since a LEN of
255 no longer indicates dynamic-ness).