Switch strings to have WIDE_BYTE() free
Following the lead from IS_ARRAY meaning SER_WIDE() returns REBVAL,
this changs it so that IS_STRING makes SER_WIDE() return 1.

This doesn't cause a lot of cost, because most code dealing with
arrays or strings already have the size hardcoded when doing accesses.
Only very generic series routines (for calculating size used, sliding
elements around) actually need the SER_WIDE().  Though it is used
in the debug builds as a double check that the right width is passed
in by callsites.