Introduce "inode" concept as alternative to "info"
R3-Alpha's series nodes were originally 5 pointers long.  Ren-C shifted
to 8...with a move toward being able to allow small series to live
directly in the nodes.  This gave more breathing room for flags (which
were in short supply) as one of the new pointers was mapped out for
flags and information bytes...while there were two slots instead of
just one for extra shared state.

Flags came to be used more cleverly, and with that cleverness opened
up the opportunity to make some series types choose to use their
second flags entry as another platform pointer.  This lays the
groundwork for that choice, calling the extra GC-markable node slot
the INODE()...alongside LINK() and MISC().

Only specific internal series types will offer the inode, recognized
by flag patterns (there are scarce flags for indicating it) and the
bonus will always be a GC-marked node for these types.

(No such types are introduced in this commit, it's just groundwork.)