Free "bias" for use in always-dynamic series
Most array and string series can add in an offset to their data so that
a TAKE from the head need not move all the memory at that time.  This
permits a later INSERT at the head to use the space.

It's a design choice that raises some questions about when to reclaim
this space--those questions weren't answered in R3-Alpha.  Only a CLEAR
of a series would reclaim the bias area.  But this fits into a lot
of unanswered questions about memory reclamation in series, which the
GC was not set up to do.

Either way--the feature can't (or shouldn't) be used by contexts or
actions, as you can't remove content from their head positions (the
head is supposed to stay where it is).  This makes it so those types
can use the slot for other purposes.  The name is BONUS, which may
help as a reminder that it's related to the BIAS.  (It's the "bonus"
you get as a reward for not using the bias...?)