Use BONUS for patchlist of context
Due to runnning short on context pointer slots, the list of patches
which had a particular context at the head had to be folded in with
the META information...with the meta information then bumped into
the patch.  Not only did this complicate the code, but it also meant
that all virtual binding patches had to give up their MISC() slot.

But since all contexts (that aren't freed) are dynamically allocated,
they now have a "bonus" slot where the "bias" would usually be.
This makes use of that slot instead.