Switch series typing from bits to "flavor" byte
Rather than have a byte taken from their header in order to indicate
a series subclass (the way values have a "kind byte"), series had been
instead using individual bits that categorized them more broadly, and
could test for aspects that spanned several series types in a single
operation...such as tests for IS_ARRAY or IS_STRING.

The categories of array subclasses grew significantly, to where it
became rather necessary to know which kind one was dealing with.  This
especially became the case as reference sites for series nodes would
make the subclass have implicit semantic meaning vs. encoding the
meaning at the callsite.

It had reached a point where the number of bits to determine the series
subclasses had grown so much that it was necessary to test one bit
before being able to test another (e.g. one had to test to see if
something was an array before being able to test for a varlist).  This
meant two bit operations for extremely common tests, which a type
byte could accomplish in one equality test.

This switches to that byte strategy--calling it a "flavor" to help
distinguish it from "type/kind" for values.  This needed big shifts
in bit positioning in the headers and info bits.

By far the largest category of subclasses is the array subclasses, and
the byte values are chosen to make a simple test of whether something
is an array by being any of the flavors less than a certain value.  A
similarly simple test checks for byte-sized series or strings.

Since the callsites were no longer reading flags, they were changed
to more readable forms, e.g.:

    IS_SER_ARRAY(s) and GET_ARRAY_FLAG(ARR(s), VARLIST)

Can be implemented in a single byte test with:

    SER_FLAVOR(s) == FLAVOR_VARLIST

Which is expressed with the lighter and clearer:

    IS_VARLIST(s)

This contributes significantly to readability.