Introduce series subclass flags
With bits previously spread out now compacted into a "flavor" byte--
as well as moving the symbol series ID number into the "INFO" area,
the previous "generic array flags" begin to make more sense as a
byte's worth of subclass-specific flags.

This establishes those subclasses, as well as adding accessors that
validate the series as being in the subclass before reading or writing
the subclass flag.  (Previously the only rigor in this area was done
that what you were dealing with was an array.)