Allow CTX() to operate on INACCESSIBLE contexts
There needs to be a more holistic model for dealing with contexts that
become INACCESSIBLE, but which may be pointed to by places that
aren't in value cells (where the extraction would fail).

This keeps what was running previously afloat by making the CTX()
macro not assret when a context is inaccessible.  However, it probably
should do so.  This way if all contexts are stored as nodes, the
conversion to a context can be the point that notices that they are
not valid for using with CTX_XXX() operations.