Fix binary copying of aliases
Because symbols can be aliased as strings, it is possible to get a
symbol series as the input to a COPY of a string operation.  When you
do, you don't want to copy the status of that series as a symbol...
because that would try to remove the symbol from the synonyms list
and do other symbol-oriented information.

A broader review of how to deal with copying "flavored" series is
needed, but this makes the immediate change to fix COPY of an
AS BINARY! of a WORD! to keep from crashing the system.  It just
basically forces the type to a binary for the problem cases.