Put DETAILS_FLAGs in subclass flags, not info
In the migration to the new "flavored" series format with subclass
flags, a first step pushed the details flags into an unused byte in
the INFO bits.  With the bits reorganized and now with an assert to
make subclass bit use safer, this goes ahead and makes the details
bits use the subclass...freeing the info byte for all non-symbol
series.