Rearrange %sys-rebxxx.h before %sys-internals.h
In trying to break up larger files into smaller ones, a natural
separation occurs between those things that have to be defined before
the %sys-internals.h header, and things that should be defined
afterwards.  In the absence of other metrics, this works to know
when to draw the line on things like flag/struct definitions vs.
things that will need APIs.

The idea of doing this was planned out for the %sys-rebser.h and
%sys-rebarr.h files, but not for others.  This does an adjustment
to align them with the rule.