Update bitfield header puns
These bits are convenient when debugging, so you don't have to reverse
engineer the binary bits in the header.  However, they do drift and
get out of date...and hadn't been synchronized for a long time.  The
recent shuffles in header bits and bytes threw them completely off,
so this at least brings them up to date for the moment.