Check subclass in debug for misc, link, bonus
Series subclasses use their LINK, MISC, and (if available) BONUS fields
in completely different ways... though they all adhere to the idea of
storing through a common node pointer that can be read as a node
without violating strict aliasing.

There are helpers which assisting in making it clearer what the data
types involved are.  However, those helpers were not doing any checking
to make sure they were being used with the right kind of series node.

Previously that would have been difficult, but the series "flavor byte"
makes this test generically easy to add.  The `ensure_flavor` macro
is coupled with the accessors to assure the right combination is used.