Add GC behavior for INFO node
This adds a flag for controlling if you use the info slot for a node,
to mark it with the GC.  Ultimately this flag may become the same flag
used to test for indirected lock information.

It also makes the subclass flag for file and line information more
generic, so that it can be used for other purposes by other subclasses
than a plain source-level array.