Change LET to a dynamic implementation
This replaces the haphazard scan for the WORD! LET done by FUNC to
add variables to `<local>` with a virtual-binding based implementation
that actually makes LET define new space for variables.

It currently uses 8 platform pointers of space per LET.  4 of those
are for the cell itself, one for the symbol pointer, one for the
series header, and then a link in the virtual binding chain as well
as the typical virtual binding patch that links to variations with
a different ->next pointer.