Fix TRANSCODE/NEXT on BINARY! with DATE!
The scanner ep state was being used, when it hadn't been reflected to
update by the TOKEN_DATE scan for the extended span of the pathlike
component.  This only happened in the BINARY! branch because the
TEXT! branch used the `bp` (and also tested it for null).  Sync the
two branches so they do the same thing.