Consolidate cell copying to one routine
The varied needs for copying a bit pattern from one cell to another
led to a mix of functions for relative, specific, or "blit" or "var"
facilitating routines.

All of the specific vs. relative discernment is for checking only,
as the C build defines them the same.  This means that C++-based
routines can use overloading to cover the permutations and return
results for what is legal and illegal (e.g. you can't copy a relative
value and get a specific one back).

The other concepts of "Var" vs "Blit" really had to be explained at
the callsite w.r.t. what they were doing with the masking of flags.
This just makes a version of the copy where you can pass the mask
along explicitly, which is more instructive and then doesn't need
to be commented.