Discern Move_Cell() and Copy_Cell()
The historical name "Move_Value()" was chosen in order to not conflict
with what might be thought of as the behavior of the COPY native,
which would act on what the value means and not the bits of the cell.

Switching to the word "Cell" helps get away from that, to where "copy"
can more obviously refer to the bits.  It also means that "move" can
be adjusted to mean truly moving the bits and mutating the old cell
to an unused state.  Over the long term, this might be leveraged in
order to do reference counting, and know that the old location no
longer is meaningful...so a series could be moved around as a "hot
potato" and freed when no references remain.

(Since there's no such trick implemented yet, the Move_Cell() variant
does nothing in the release build...but it does trash the moved from
cell via mutable access in the debug build, to start testing the
potential future feature.)