Fix LET at global scope, add starter test file
Making a LET at global scope was causing an assertion due to not
testing for NULL specifiers.  This addresses that, factors the
creation of a LET patch into its own routine, and adds a starter
test file.

(LET was committed without tests because its previous experimental
incarnation had no tests, beyond just the usages in the mezzanine.