Make EVALUATE preserve LET Bindings (dicey, but works)
This implements the idea of accumulating bindings from LET on the
BLOCK! returned by EVALUATE, thus allowing the block value alone to
represent the accrued state.

It may not be the best way to do it...as users could then navigate
the block (e.g. take its HEAD) and be at the start, yet still with
this additional element of accrued LET state.  If they evaluate step
again, they would be doing so with the bindings downstream already
in effect (which would give differing results)...and then they would
be adding even more bindings to the chain.

A better answer is likely to accrue this state in something like a
VARARGS! or a FRAME!, where more restricted navigation of the input
is in effect.  Then if someone wanted to go back to the raw block,
they'd have to do so from the original material.

https://forum.rebol.info/t/1496