Kill SYNC-INVISIBLES
Originally, the evaluator tried to be sneaky about invisible evaluation
by folding it in with other evaluations.  This meant when you thought
you were taking a single step in an evaluation, you might look ahead
and see something that wasn't what you were looking for...and then
step...and it would bundle the execution of what you were looking for
along with an invisible so you passed it.

So for something that was running an EVALUATE in steps, there had
been an idea that you could explicitly call SYNC-INVISIBLES to ask it
to bring all the invisible executions up to date.  This is no longer
feasible due to the fact that invisibility isn't necessarily a pre
announced property...a function can decide during it's run if it is
going to be invisible or not.

However, this doesn't matter since the original intention was to let
you cue up to something that you might be interested in...and the
EVALUATE primitive lets you run evaluations and see the invisibles
in steps with a specially marked "the step was invisible" output
(currently done by the QUOTED! status).