Switch all FUNCT-style FUNCTION usages to FUNC w/LET
The idea of collecting SET-WORD!s and assuming they represent local
variables in a function is problematic.  It winds up to the creation
of locals not just for a function, but also for any functions nested
inside of it... or any object keys, or any SET-WORD!s that are used
in dialects for other purposes.

This fostered the idea of being able to do a dynamic form of variable
creation with LET.  While this is more expensive in terms of memory,
CPU, and GC load...it offers more power, and isn't the dead-end of
the locals-gathering via SET-WORD!.

If it proves to be too expensive for certain tasks, then manual locals
that are pre-declared via `<local>` can be used instead.