Catch bad UTF-8 during WORD! Scanning
R3-Alpha's scanner split the process of scanning into "finding a token"
(where it started and ended based on spaces and delimiters) and then
passing that start and end point to a function whose job it was to
make that specific type (WORD!, etc.)

The boundary-finding for words did not do UTF-8 validation, and the
second phase of interning the string data to make the word also did
not check it.  This meant bad UTF-8 words could be imported, which
the system was not prepared to handle.

There actually were two different hash routines for unchecked and
checked data, but they didn't use the same hash.  This converges the
routines to use the crc32 hash (a sunk cost due to zlib) and makes
their data types reflect whether they take already-checked UTF-8
data or not.

With everything using the zlib hash, the CRC24 isn't necessary in the
core.  Since the usefulness of CRC24 is questionable at this point,
it's moved to the crypt module as unused...with a failure raised if
someone tries to call it.

Also does some touch ups to the DEBUG_UTF8_EVERYWHERE build in the
process.