Rename EQUAL-OR-LESSER? to LESSER-OR-EQUAL?
There was an idea tried out for a while to get =< and >= to be the
operators for comparisons.  This way, => and <= could be seen as
"arrows".

While this idea was pushed, the =< written out phrase was as
"EQUAL-OR-LESSER?"...because that was supposed to be the "real operator"
and free up the <= to be an arrow in the language.

Given many people's staunch feeling that the "less than or equalness"
of `<=` was intrinsic and obvious, that suggests that `=>` is also
obviously "equal or greater".  For this (and other reasons) => was
removed as an "in the box" operator distinct from >=.

With -> now acting as lamda, this goes ahead and fleshes out the
compromise to define => as equal or greater, and to give synonyms in
the written out forms as well (LESSER-OR-EQUAL? is <=)