Annotate FILE and LINE on errors prior to PROBE()
The debug option that probes every error that goes by is important
in the JavaScript build in particular.  However, the probe was being
done before the file and line were added to the error (if it did
not have one already, and it needed fail() itself to add it)

This reorders the code to earlier annotate a fail() on an undecorated
error (e.g. one that originates from inside the interpreter, not FAIL
the usermode function).