Get rid of SYS/INIT-SCHEMES, simplify DECODE-URL
Over time, Ren-C has pushed the "scheme" definitions so that none of
them are "baked in"... so it's possible to build a core which has
no schemes hardcoded into it.

This goal was accomplished, but the INIT-SCHEMES routine still existed
from when it was used to register "core" schemes.  This removes that.

Also, there were a couple of things that made DECODE-URL confusing.
One is that it is defined in `sys` because it is used by the
general MAKE-SCHEME mechanics, and it needed a separate definition
to get added to lib to be called as plain `decode-url` instead of
`sys/decode-url`.

Another aspect that made it confusing was that it used several rules
bound into a context.  These rules didn't want to be top level
definitions.  So the DECODE-URL function was put in that context
as well, and then had to be exposed.

It was more complex than it needed to be, so this makes it more obvious
in line with the removal of INIT-SCHEMES.