Fix the fix for file/line in erros
It was initially thought that the problem with file/line reporting
was that it wasn't climbing the frames, but it already was...the
problem was that the error probe was happening before that climb.

The added additional climbing was not removed from the commit fixing
the actual problem, and unlike the version that already existed it
had a bug in it by using the file information before checking the
flag for whether it was present.

This removes the added buggy climbing.